\name{osc1337-package}
\alias{osc1337-package}
\docType{package}
\title{
  Terminal Graphics using OSC-1337
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Neal Fultz, Mike Frysinger for original shell scripts

Maintainer: Neal Fultz <nfultz@gmail.com>
}
\references{
  The inline image protocol is defined at \url{https://www.iterm2.com/documentation-images.html}.
}
\keyword{ package }
\examples{
  \dontrun{
      library(osc1337)
      barplot(1:10) # should appear in terminal.
  }
}
